refactor(image_build_manager): split build_image_func, fix cred sync, rename suffixes
- Fix credential sync: replace heredoc with printf (paramiko compat)
- Rename build type suffixes: -ib/-th to -imgbld/-imgth
- Fix naming tests (TC_BD_007-011): use curl API instead of broken
  regctl/s3cmd approach for registry and S3 checks
- Split build_image_func.py (1963 lines) into 9 focused modules:
  _config_helpers, container_func, s3_func, registry_func,
  build_status_func, image_content_func, cleanup_func,
  precheck_func, log_func
- build_image_func.py retained as re-export hub (backward compat)
- Add shell quote escaping for credential values
