+++ test/image_build_manager/library/vars/test_case_vars.py
@@ -115,23 +115,23 @@ TEST_CASES = {
     # ── Build — naming convention (TC_BD_007-011) ─────────────────────────
     "registry_naming_ib_x86_64": {
         "id": "TC_BD_007",
-        "title": "Verify image-builder registry naming (-ib suffix, x86_64)",
+        "title": "Verify image-builder registry naming (-imgbld suffix, x86_64)",
     },
     "s3_naming_ib_x86_64": {
         "id": "TC_BD_008",
-        "title": "Verify image-builder S3 naming (-ib suffix, x86_64)",
+        "title": "Verify image-builder S3 naming (-imgbld suffix, x86_64)",
     },
     "registry_naming_th_x86_64": {
         "id": "TC_BD_009",
-        "title": "Verify image-thrillhouse registry naming (-th suffix, x86_64)",
+        "title": "Verify image-thrillhouse registry naming (-imgth suffix, x86_64)",
     },
     "s3_naming_th_x86_64": {
         "id": "TC_BD_010",
-        "title": "Verify image-thrillhouse S3 naming (-th suffix, x86_64)",
+        "title": "Verify image-thrillhouse S3 naming (-imgth suffix, x86_64)",
     },
     "artifact_suffix_isolation": {
         "id": "TC_BD_011",
-        "title": "Verify -ib and -th artifact paths are fully isolated",
+        "title": "Verify -imgbld and -imgth artifact paths are fully isolated",
     },
 
     # ── Full (image_build_manager) ────────────────────────────────────────
